+++ compiler/src/jdk.internal.vm.compiler/src/org/graalvm/compiler/asm/amd64/AMD64Assembler.java
@@ -2948,7 +2948,7 @@ public class AMD64Assembler extends AMD64BaseAssembler {
         emitModRM(4, entry);
     }
 
-    public final void jmp(Register entry) {
+    public void jmp(Register entry) {
         int bytesToEmit = needsRex(entry) ? 3 : 2;
         mitigateJCCErratum(bytesToEmit);
         int beforeJmp = position();
@@ -2956,7 +2956,7 @@ public class AMD64Assembler extends AMD64BaseAssembler {
         assert beforeJmp + bytesToEmit == position();
     }
 
-    public final void jmp(AMD64Address adr) {
+    public void jmp(AMD64Address adr) {
         int bytesToEmit = getPrefixInBytes(DWORD, adr) + OPCODE_IN_BYTES + addressInBytes(adr);
         mitigateJCCErratum(bytesToEmit);
         int beforeJmp = position();
@@ -4071,7 +4071,7 @@ public class AMD64Assembler extends AMD64BaseAssembler {
         emitModRM(dst, src);
     }
 
-    public final void ret(int imm16) {
+    public void ret(int imm16) {
         if (imm16 == 0) {
             mitigateJCCErratum(1);
             emitByte(0xC3);
@@ -4851,7 +4851,7 @@ public class AMD64Assembler extends AMD64BaseAssembler {
      * all calls need patching anyway. Therefore, 0 is emitted as the call target, and the user is
      * responsible to add the call address to the appropriate patching tables.
      */
-    public final void call() {
+    protected final void call() {
         annotatePatchingImmediate(1, 4);
         emitByte(0xE8);
         emitInt(0);