+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/heap/ReferenceAccessFeature.java
@@ -1,11 +1,10 @@
 /*
- * Copyright (c) 2023, 2023, Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2023, 2023, BELLSOFT. All rights reserved.
+ * Copyright (c) 2024, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation. Oracle designates this
+ * published by the Free Software Foundation.  Oracle designates this
  * particular file as subject to the "Classpath" exception as provided
  * by Oracle in the LICENSE file that accompanied this code.
  *
@@ -23,22 +22,17 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
+package com.oracle.svm.core.heap;
 
-#ifdef __linux__
+import org.graalvm.nativeimage.ImageSingletons;
 
-#include <sys/syscall.h>
-#include <sys/types.h>
-#include <unistd.h>
+import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
+import com.oracle.svm.core.feature.InternalFeature;
 
-/*
- * Based on os::Linux::gettid() from jdk-20-ga, see
- * https://github.com/openjdk/jdk20/blob/82749901b1497f524e53e47c45708c8e4a63c8b9/src/hotspot/os/linux/os_linux.cpp#L361
- *
- * syscall() uses varargs, so we can't use @CFunction.
- */
-pid_t getThreadId() {
-  return (pid_t)syscall(SYS_gettid);
+@AutomaticallyRegisteredFeature
+public class ReferenceAccessFeature implements InternalFeature {
+    @Override
+    public void afterRegistration(AfterRegistrationAccess access) {
+        ImageSingletons.add(ReferenceAccess.class, new ReferenceAccessImpl());
+    }
 }
-
-#endif
-