+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/code/amd64/AMD64HostedPatcherFeature.java
@@ -40,19 +40,16 @@ import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.core.graal.code.CGlobalDataReference;
 import com.oracle.svm.core.graal.code.PatchConsumerFactory;
-import com.oracle.svm.core.meta.MethodPointer;
 import com.oracle.svm.core.meta.SubstrateMethodPointerConstant;
 import com.oracle.svm.core.util.VMError;
 import com.oracle.svm.hosted.code.HostedImageHeapConstantPatch;
 import com.oracle.svm.hosted.code.HostedPatcher;
 import com.oracle.svm.hosted.image.RelocatableBuffer;
-import com.oracle.svm.hosted.meta.HostedMethod;
 
 import jdk.vm.ci.code.site.ConstantReference;
 import jdk.vm.ci.code.site.DataSectionReference;
 import jdk.vm.ci.code.site.Reference;
 import jdk.vm.ci.meta.JavaConstant;
-import jdk.vm.ci.meta.VMConstant;
 
 @AutomaticallyRegisteredFeature
 @Platforms({Platform.AMD64.class})
@@ -122,17 +119,9 @@ class AMD64HostedPatcher extends CompilationResult.CodeAnnotation implements Hos
              */
             long addend = (annotation.nextInstructionPosition - annotation.operandPosition);
             relocs.addRelocationWithAddend((int) siteOffset, ObjectFile.RelocationKind.getPCRelative(annotation.operandSize), addend, ref);
-        } else if (ref instanceof ConstantReference) {
-            VMConstant constant = ((ConstantReference) ref).getConstant();
-            Object relocVal = ref;
-            if (constant instanceof SubstrateMethodPointerConstant) {
-                VMError.guarantee(!Platform.includedIn(Platform.DARWIN_AMD64.class), "[GR-43389] method pointer relocations should not be inlined.");
-                MethodPointer pointer = ((SubstrateMethodPointerConstant) constant).pointer();
-                HostedMethod hMethod = (HostedMethod) pointer.getMethod();
-                VMError.guarantee(hMethod.isCompiled(), "Method %s is not compiled although there is a method pointer constant created for it.", hMethod);
-                relocVal = pointer;
-            }
-            relocs.addRelocationWithoutAddend((int) siteOffset, ObjectFile.RelocationKind.getDirect(annotation.operandSize), relocVal);
+        } else if (ref instanceof ConstantReference constantRef) {
+            VMError.guarantee(!(constantRef.getConstant() instanceof SubstrateMethodPointerConstant), "SubstrateMethodPointerConstants should not be relocated %s", constantRef);
+            relocs.addRelocationWithoutAddend((int) siteOffset, ObjectFile.RelocationKind.getDirect(annotation.operandSize), ref);
         } else {
             throw VMError.shouldNotReachHere("Unknown type of reference in code");
         }