+++ truffle/mx.truffle/mx_truffle.py
@@ -1027,6 +1027,8 @@ class LibffiBuilderProject(mx.AbstractNativeProject, mx_native.NativeDependency)
         3. Invoking the platform dependent builder that we delegate to.
     """
 
+    libs = property(lambda self: self.delegate.libs)
+
     def __init__(self, suite, name, deps, workingSets, **kwargs):
         subDir = 'src'
         srcDirs = ['patches']
@@ -1088,7 +1090,6 @@ class LibffiBuilderProject(mx.AbstractNativeProject, mx_native.NativeDependency)
             )
 
         self.include_dirs = self.delegate.include_dirs
-        self.libs = self.delegate.libs
 
     def resolveDeps(self):
         super(LibffiBuilderProject, self).resolveDeps()